trainingtools/requirements.txt
Pillow
python-resize-image
pandas